Added web page configuration flexibility.
Modified the CSS templates to specify needed points of flexibility for
ESnet skinning of Safnari. Placed all modified values in variables.less
so CSS tailoring can be reduced to the single file.  Also added web
page footer configuration to application.conf file so per deployment
contact information can be specified for display in web page footer
(not sure if this was the best way but it does the job for now).  All
defaults give the SURFnet look and feel.