Use one of the child errors to report to requester
When one or more child NSA fail we pick one of them as "prioritized"
error and send that the to requesting NSA. The child exceptions are
populated with the original child errors.

Right now the prioritization is somewhat random (the "first" from a hash
map). Better prioritization may be done using error codes or some other
strategy.