Support `ifModifiedSince` in NSI 2.1 queries
Also return `LastModified` in every query response (as mandatory part of the
`ifModifiedSince` spec).

Track the last updated timestamp on a `ConnectionEntity` so that we actually
know the `LastModified` per connection. While replaying use a fixed clock so
that the original update timestamps are tracked.

Also use a fixed clock during any transaction so that inbound and outbound
messages share the same "transaction" timestamps.