queryRecursiveConfirm message fix
nsi-safnari was not populating the p2ps element associated with its service entry in the criteria element of the queryRecursiveConfirmed.  This p2ps element was held in an Any structure that was not being copied into the queryRecursiveConfirm from the original reservation.   Added copt and issue is fixed.  Closed issue #18.